+++ docs/changes.rst
@@ -17,6 +17,7 @@ Release history
 - feat: digit patterns APi in outgoing calling permissions for locations, users, workspaces, and virtual lines, :attr:`wxc_sdk.person_settings.permissions_out.OutgoingPermissionsApi.digit_patterns`.
 - feat: first (experimental, rudimentary) shot at SCIMv2 users API :attr:`wxc_sdk.scim.users.SCIM2UsersApi`, only implemented :meth:`wxc_sdk.scim.users.SCIM2UsersApi.details` and :meth:`wxc_sdk.scim.users.SCIM2UsersApi.search`
 - feat: :meth:`wxc_sdk.scim.users.SCIM2UsersApi.search_all`, :meth:`wxc_sdk.scim.users.SCIM2UsersApi.update`, :meth:`wxc_sdk.scim.users.SCIM2UsersApi.patch`, :meth:`wxc_sdk.scim.users.SCIM2UsersApi.delete`
+- feat: SCIMv2 bulk API :attr:`wxc_sdk.scim.bulk.SCIM2BulkApi`
 
 1.17.1
 ------